// src/app/model/credit-product.ts
export class CreditProduct {
  sId?: number;
  sName?: string;
  sDefProdId?: number;
  sPartPay?: number;
  sGracePeriod?: number;
  maxTerm?: number;
  firstPay?: number;
  monthCommissionPct?: number;
  yearPct?: number;
  kpcPct?: number;
  minAmt?: number;
  maxAmt?: number;
  minTerm?: number;

  constructor(){};
}